requirements.txt
asgiref==3.8.1
boto3==1.36.1
botocore==1.36.1
Django==5.1.3
django-storages==1.14.4
jmespath==1.0.1
pillow==11.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.11.1
six==1.17.0
sqlparse==0.5.2
tzdata==2024.2
urllib3==2.3.0
